Move Server to nodelib and add --host server option
